+++ airvpn/web/user/__init__.py
@@ -63,9 +63,9 @@ class WebUser:
         self._session = session
         self.name = kwargs.get("name")
         self.id = kwargs.get("id")
-        self.image = kwargs.get("image")
         self.profile_url = f"{WebSession.__BASE_URL__}/profile/{self.id}-{slugify_username(self.name)}/"
 
+        self._image = kwargs.get("image")
         self._content_count = kwargs.get("content_count")
         self._followers = kwargs.get("followers")
         self._community_reputation = kwargs.get("community_reputation")
@@ -79,6 +79,56 @@ class WebUser:
         self._contacts = kwargs.get("contacts")
 
         self._cache_ts = 0
+
+    def follow(self) -> bool:
+        """Follow this member.
+
+        Returns:
+            bool: ``True`` if the request succeeded.
+        """
+        if not self._session.logged_in:
+            return False
+
+        response = self._session.ajax("post", "follow", "notifications", ajax_params = {
+            "follow_app": "core",
+            "follow_area": "member",
+            "follow_id": self.id
+            },
+            files=(
+                ("follow_submitted", (None, 1)),
+                ("csrfKey", (None, self._session.csrf)),
+                ("immediate", (None, "follow_type_immediate")),
+                ("follow_public", (None, 0)),
+                ("follow_public_checkbox", (None, 1))
+            ))
+
+        status_code = response.status_code
+        return status_code == 200 or status_code == 301
+
+    def unfollow(self) -> bool:
+        """Unfollow the user.
+
+        Returns:
+            bool: ``True`` if the request succeeded.
+        """
+        if not self._session.logged_in:
+            return False
+
+        response = self._session.ajax("get", "follow", "notifications", 
+            ajax_params={
+                "follow_area": "member",
+                "follow_id": self.id,
+                "follow_app": "core"
+            })
+
+        soup = BeautifulSoup(response.text, "html.parser")
+
+        following_member = soup.find("a", {"data-action": "unfollow"})
+        if following_member is None:
+            return False
+
+        response = self._session.session.get(following_member.get("href"))
+        return response.status_code == 200 or response.status_code == 301
 
     def _cache_profile(self):
         """Fetch and parse the user's profile page, populating all lazily
@@ -113,7 +163,7 @@ class WebUser:
 
         self.name = name_elm.text.strip()
         self._rank = rank_elm.text.strip()
-        self.image = img_elm.get("href")
+        self._image = img_elm.get("href")
 
         profile_stats = soup.find("div", id="elProfileStats")
 
@@ -211,6 +261,12 @@ class WebUser:
             """Force update attributes."""
             self._cache_ts = 0
             self._cache_profile()
+
+    @property
+    def image(self):
+        """str: The user's profile picture."""
+        self._cache_profile()
+        return self._image
 
     @property
     def contacts(self):